Fix timing readout in cal_points segments
The timing of the readout pulse wasn't coming after the last pulse (in
time unit) but after the end of the last pulse of the pulse list.
This typically results in errors when some qubits are prepared in the e
state and others in the f state.
Corrected by using the multireferencing such that the readout pulses
would start after the last state preparation pulse