Add parallel SSRO analysis
- based on MultiqubitTimedomain: while this seems strange at first,
the required "per qubit" handling of the data is quite similar.
I could not include the analysis in readout_analysis because it would
create a circular dependency between timedomain_analysis.py and readout_analysis.py
When migrating to analysis_v3, one should be able to just reuse common nodes.
For now, it resulted in less code duplication to do it in that way.
- Reuses all static functions from the SSRO Qutrit single shot analysis
- Basic functionalities implemented and tested on virtual setup