Fix encoding when saving to hdf5
keys for attributes must be of string type. If it is not the case (e.g. the key of dictionary is a integer) then this results in an Exception and corresponding key/value pair is not saved.
To fix this, a representation of the key is created. To retrieve the original value, "eval"
must be used on the representation, i.e. eval(repr(key)) == key