Fix state-to-integer mapping to allow arbitrary state order
The Gaussian Mixture model predicts labels using
range(n_gaussian_components), because it is an unsupervised model.
So when a qubit had only two states, e.g. "g" and "f", it would
predict 0 and 1's but prep_state assumed f= 2. This can cause issues
e.g. when states appearing in cal states are not in the default order
"g", "e", "f". Because the init_means passed to the GMM serves as a
guideline to guess which state will have with label output from the
model (since the model is unsupervised, there is no other
straightforward way to tell the model which component should be 0, 1 or
2). If e.g. "f" occurs first in cal_points, then it would be passed as
  the first "init_means" and thus would be predicted as 0.
  To keep flexibility about state labels and allow for states to be
  added later (e.g. "h" state or maybe multiqubit states), I didn't
  create a manual reordering but instead I adapt the mapping according
  to how states occur in cal_points, for each qubit separately.