Fix ordering of states in state prob assignment matrix
Allow for states to be passed in arbitrary order, while
still plotting them in the "usual order" in the state assignment
probability matrix. e.g. when measuring "e", then "g",
the order in the matrix will still be "g" then "e". If the state
labels are not known, then the order in which they are measured
is taken.
Fix bug related to ordering: pass the correct state labels integers to
the confusion_matrix function instead of the "order" of the labels (i.e.
indices)